backend/pyproject.toml
[project]
name = "cimo"
version = "0.1.0"
description = "Web application for personalized movie selection"
authors = [
    {name = "Sergei Vdovenkov", email = "vdovenkov2000@gmail.com"},
    {name = "Ivan Golyshev", email = "golysheev@gmail.com"},
    {name = "Nikita Petrov", email = "nikita.petrov.it@gmail.com"},
    {name = "Elena Miacheva", email = "mmiacheva@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "sqlalchemy (>=2.0.36,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic[email] (>=2.11.1,<3.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "scikit-learn (>=1.6.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.1,<3.0.0)",
    "joblib (>=1.4.2,<2.0.0)",
    "datasets (>=3.2.0,<4.0.0)",
    "torch (>=2.5.1,<3.0.0)",
    "pytorch-lightning (>=2.5.0.post0,<3.0.0)",
    "transformers (>=4.47.1,<5.0.0)",
    "torchvision (>=0.20.1,<0.21.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "sqladmin[full] (>=0.20.1,<0.21.0)",
    "bcrypt (==4.0.1)",
    "fastapi-cache2[redis] (>=0.2.2,<0.3.0)",
    "celery (>=5.4.0,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "faker (>=37.1.0,<38.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^25.1.0"
pytest-cov = "^6.0.0"
mypy = "^1.15.0"
isort = "^6.0.0"
httpx = "^0.28.1"
ruff = "^0.9.5"

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["app"]

[tool.mypy]
strict = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
explicit_package_bases = true

[tool.black]
line-length = 120 
target-version = ['py311']

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "C"]
